Create Main
I want to make a collision between the monkey and the cage, but i can`t. I can't find anything. I tried to make a collision detection but it dose nothing and even if it would work, I think that it will return true only if the center of the monkey and the center of the cage are at the same coordinates.